[CL-2102] Integrate & remove nav_bar_items_controller & routes patches
