--- back/engines/free/customizable_navbar/config/routes.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-CustomizableNavbar::Engine.routes.draw do
-  namespace :web_api, defaults: { format: :json } do
-    namespace :v1 do
-      resources :nav_bar_items, only: %i[create update destroy] do
-        patch 'reorder', on: :member
-      end
-    end
-  end
-end
-
-Rails.application.routes.draw do
-  mount CustomizableNavbar::Engine => ''
-end
